Se añadieron carpetas y helpers - assets - helpers - context - hooks Cambio a archivos .jsx
