+++ client/src/data/TickerList.js
